feat(static-analysis): Q3 psalm ladder and scope parity
Register psalm/plugin-laravel (v3.0.x, the only line compatible with the
OSPanel-managed PHP 8.4.1 runtime and the pinned Psalm 6.4.1) and extend
psalm scope to routes/ for PHPStan parity. Raise errorLevel 6 -> 5 -> 4
by source-typing fixes (no baseline accumulation): AppServiceProvider
migrated isProduction() to environment() (Application contract); redundant
casts removed in webhook/maintenance/reconcile paths; now()->timestamp
replaced with getTimestamp() (interface-declared accessor); Promotion
model annotated with full @property inventory.

Documented plugin-version tradeoff: TooManyTemplateParams (4 dirs/files)
and InvalidDocblock (app/Models + app/Repositories) suppressed via
psalm.xml issueHandlers with enumerable scope, because psalm/plugin-laravel
v3.0.x types LengthAwarePaginator and Eloquent relations as 1-param while
larastan types them as 2-param; upstream fix lands in plugin v3.14
(PR #1082, #1141) which requires PHP ~8.4.3 + Psalm ^6.16.1, blocked by
the current OSPanel PHP runtime. Suppressions are removable in a single
edit once the runtime is upgraded (risk register #22).

PsalmLadderScopeParityGuardrailTest (8 assertions) locks errorLevel <= 4,
the extended scope, plugin registration, baseline-free progression,
findUnusedBaselineEntry=true, documented template-arity suppressions,
composer constraint window, and the environment() contract.

+++ app/Support/Orders/OrdersReconcileRunner.php
@@ -58,7 +58,7 @@ final class OrdersReconcileRunner
 
     private function emitTelemetry(OrdersReconcileRunResultDto $result): void
     {
-        $failedJobsCount = $result->failedJobs === [] ? 0 : (int) ($result->failedJobs[0]->count ?? 0);
+        $failedJobsCount = $result->failedJobs === [] ? 0 : ($result->failedJobs[0]->count ?? 0);
 
         Log::info('observability.reconciliation', [
             'clean' => $result->isClean(),
